remove un-used filters
